Move test helper to helpers file
