+++ tests/helpers.py
@@ -2,6 +2,11 @@ from nexus_constructor.nexus.nexus_wrapper import NexusWrapper
 from nexus_constructor.component import Component
 from uuid import uuid1
 from typing import Any
+import h5py
+
+
+def create_in_memory_file(filename):
+    return h5py.File(filename, mode="x", driver="core", backing_store=False)
 
 
 def create_nexus_wrapper() -> NexusWrapper:
